# docs/api/openapi.json
{
  "components": {
    "responses": {
...
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        },
        "description": "400 error."
      },
      "Channel": {
        "content": {
          "application/json": {
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          }
        },
        "description": "Channel response."
      },
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        },
        "description": "403 error."
      },
...
      "Guild": {
        "content": {
          "application/json": {
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          }
        },
        "description": "Guild response."
      },
...
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        },
        "description": "404 error."
      },
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        },
        "description": "401 error."
      },
...
    },
    "schemas": {
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "requestId",
          "code",
          "message",
          "status"
        ],
        "type": "object"
      },
      "CreateChannelRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "parentId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "type": "object"
      },
      "CreateGuildRequest": {
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Discord Clone API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/guilds": {
      "post": {
        "operationId": "createGuild",
        "parameters": [
          {
            "description": "Safe request correlation id. Unsafe incoming values are replaced by the backend.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Bearer access token where authentication is required.",
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Guild"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Create a guild.",
        "tags": [
          "guilds"
        ]
      }
    },
    "/api/guilds/{guildId}/channels": {
      "post": {
        "operationId": "createChannel",
        "parameters": [
          {
            "description": "Safe request correlation id. Unsafe incoming values are replaced by the backend.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Bearer access token where authentication is required.",
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "guildId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Channel"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Create a guild channel.",
        "tags": [
          "guilds"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
...
}